rename com.jetbrains.kmpapp -> org.example.app, KotlinProject -> NewProjectName
