+++ .fleet/receipt.json
@@ -15,5 +15,5 @@
             }
         }
     },
-    "timestamp": "2024-06-20T15:08:25.549849307Z"
+    "timestamp": "2024-06-25T13:33:05.098266152Z"
 }
