+++ composeApp/src/androidMain/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">KotlinProject</string>
+    <string name="app_name">NewProjectName</string>
 </resources>
