+++ iosApp/Configuration/Config.xcconfig
@@ -1,3 +1,3 @@
 TEAM_ID=
-BUNDLE_ID=com.jetbrains.kmpapp.KotlinProject
-APP_NAME=KotlinProject
+BUNDLE_ID=org.example.app.NewProjectName
+APP_NAME=NewProjectName
