+++ settings.gradle.kts
@@ -1,4 +1,4 @@
-rootProject.name = "KotlinProject"
+rootProject.name = "NewProjectName"
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 pluginManagement {
